Get a basic cross-platform Linux / Mac build working (#1)
We've changed around the meson build so that the needed libraries will 
be built locally, even if versions of the dependencies are available
on the system.

Currently, we attempt to build statically on Linux, but not on the Mac.

We ended up not being able to use most of the out-of-the-box Meson
wraps, and had to get pretty kludgy to handle the fact that libcurl
depends on OpenSSL, since Meson tries to run autoconf on both before
building either, but libcurl's autoconf tests for the presence of the
OpenSSL .a files.

We have more work to do here for sure. We'll need to get it to build
with MUSL on Linux as an option. And on the Mac, it does not
automatically copy dylibs where they need to be.

But miracuolously, it does work on both OSs at once, though if
building outside of Docker, some distros still are going to have a
problem with libatomic.

Still, this is all good enough for now.

Miro did the hard work here :)