Didn't have as much time as I expected today, but:
1. terminal_io.c: This is just the wrapping of stdin, stdout, stderr
   fds into objects. This is all akin to the functions n00b_std*() in
   the old stream interface, and these functions will end up w/ those
   names when it's time to fully nuke the old interface.

2. Added some fns to the duration type to convert milliseconds to a
   duration object (which is really a struct timeval). This was
   because other APIs we're using require the int ms, and I needed to
   pass it to nanosleep in some places.

3. Fixed some minor bugs, like forgetting to make the fcntl call to
   get fd flags in the event system, but then using those bits I
   didn't get to determine the r/w perms on the descriptor.

4. Added in a replacement debugging interface based on the new IO
   system, but with an initial interface for per-topic filtering /
   workflow logic.

   The default workflow (if one isn't available specifically for a
   topic) is:

   - If there's a debug server available, log to it and stop. This
     checks the N00B_DEBUG_HOST and N00B_DEBUG_PORT env vars, and
     tries the default values if not available ("127.0.0.1" and
     7878). Note that currently, the HOST value must be an IP address,
     not a DNS name. But v6 addrs should work.

   - Next, IF there's a N00B_DEBUG_LOG env var set, it tries to open
     that file. If that works, it's one.

   - Finally, if neither of those worked, it writes to stderr.

   I think the approach will need a bit of tweaking. For instance, I
   added a notion of a 'formatting function'; at any point a workflow
   can change its formatting function. And, there's a default (see the
   test program for an example).

   The problem is that the default function uses ANSI codes, but we
   probably don't want that for writing to a log file. So we'll need
   to either have 2 defaults, or by default strip ansi from files, or
   something like that.

Also, I was going to whip back up the debug server before I went to
bed, but it occured to me that one critical piece of how it works is
by marshalling messages on the client side, and unmarshalling them on
the server side.

SO, it turns out the filtering needs to be done before I can get to
the rewrite of the debug server.

After I board my plane tomorrow, I'm probably going to:

1. Build a 'proxy' channel type. I wasn't sure I was going to need
them w/ the new design, but it occurs to me it's still valuable so
that you can have different stream for different filter stacks.  So I
can have 'stdout' be unfiltered, but have one proxy that ansifies
strings, and another that converts data to hex for debugging.

As an aside, (and not for tomorrow) It might even make sense to have
the functions like n00b_stdout() cache a proxy object; any time it
gets called, if it sees that filters got added in the cached object,
it creates a new proxy to return.  Though, in practice, such filters
are more likely to sit in front of debug topics really. Curious about
thoughts on that one.

2. Add temporary API to replatform the n00b_print interface.

3. Start the same replatforming on the core `proc` module (subprocess)

I think the replatformings might be easy; they leverage the old stream
interface, but I think it's not doing much more than creating and
subscribing stuff, and won't be too hard.

Whenever that's all done, I'll move to the session API, which is a
significant piece of code, but also probably doesn't require too much
work in the grand scheme of things.

After that, there are a few nits to go back and look at before
swapping out the whole IO system. For instance, there are a couple of
IO bits that I don't think are used anywhere that I haven't hit, and a
few bits that need to stick around but I don't think use the stream
interface, and I need to make sure they don't.

Once that's done, if the filtering is also done, it should probably be
time to make the switch. There may be a couple of areas that require
small bits of work, like the old system had some memory management
implications (the new one does not).

Though, there will still be at least couple of little clean-up items:

1. I'll at some point to add a 'signal' channel to replace the old
   stream versio (which was backended by libevent). I didn't rush on
   that, because it's got some minor complexities, and it's easy in
   the short term to just skip signals being implemented via channels;
   I don't really take advantage of it, beyond using one signal
   handler to publish to many registered callbacks, which only gets
   used for SIGWINCH handling right now.

2. For the debugging, I want to add the ability to toggle (via env
   var) a 'default allow' and 'default deny' policy per-topic, and
   also an exemption list. Then, I can add a test at the front to
   check these things. You can, of course, circumvent w/ a custom
   workflow, but will remove the need to use the custom workflows most
   of the time, I think.

3. The initial debug log server. Probably will also give it its own
   default workflows, at least w/ different env vars. Eventually, this
   is something I'm going to want people to be able to customize via
   config file (and thus directly in n00b). But I'll kick that well
   down the road, until I work on the REPL.