feat: Porting of proc to the new IO system
Ported the 'proc' API to the new channels, fixed a couple of minor
bugs, even though one took me way too long too long to find-- FD
caching wasn't throwing away closed fds properly, so when the OS
reused a descriptor number, stuff was wonky. I spent a lot of time
thinking I was calling open() wrong because my write-only log file
looked like it was opening read-only somehow). I also had not quite
managed to get the semantics I had intended for 'blocking' reads, and
had some more clarity of thought (this is for the low-level interface
that is char data only, not the channel interface which doesn't accept
lengths).

Now, if you pass -1 to the length field of the lower-level blocking
read call, it says, "Wait for data to become available, then read
everything available, but don't keep waiting".

I also finally got around to testing the buffer channel, so now All
the basics the basics are working. And I did started minor cleanup
work (literally-- focused on making sure unused data gets cleaned up
automatically where reasonable.

There's now a central stream registry in the constructor to aid with
debugging. Currently, it doesn't discard closed channels, but there's a
 variable controlling that, and I will default it to 'off' later.

To support subprocess management, fleshed out the rest of the core IO
objects it uses, like proxy.

Also added minor missing bits from the old API in:

- One-shot reading of file contents into either buffers or strings
  (`n00b_read_file()`)
- All the print functions now use the new IO.

Additionally: Added to the lock-free dictionary a variant version of
replacement that only replaces if the value is as expected. I only
added it to the underlying crown algorithm and the *dict* wrapper of
it, not to the 800 other algorithms.