feat: new signal handling event API
Added signal handling replacement API for the libevent signal handling
infrastructure that I was using.  Unlike them, this actually passes
the `siginfo_t` structure, not just the signal number.

The signal handler thread does not dispatch in the middle of the
signal; it flags the main loop, which calls handlers, so doesn't have
any async-signal problems.

The signal handling thread is separate from the main IO event thread,
and it does block indefinitely until a signal comes in (unlike the
main thread, which polls using a timeout).

We allow the signal handler API to take a user-suppled stashed value;
this must also be provided when unregistering; both the callback and
stash must match.