Added a string channel (though I haven't hooked it up yet).
Then I proceeded to make an awful lot of changes in the process of
porting to Linux, to try to figure out the magic problem I was having,
where things were working in a debugger, but not outside it.

Most of them seemed to not be necessary. It looks like the problem
was, I somehow started to get a poll trigger for the stdout fd to be
readable before the stdin trigger.

I don't know why my machine suddenly freaked out, but the issue does
make sense. Stdin and stdout are the same underlying pty when I'm
running at the command line, but the debugger has them as two pipes,
and not a terminal.

Some of the work I did was useful, or made things simpler: For instance:

1) I moved the 'stop the world' used in Garbage Collection from a spin
   lock to use progressive locking.

2) I added an option to protect the entire heap, and align allocs to
   the page, so if there's a pointer math error, I can detect it as it
   happens.