feat: Filters
Implemented the new filtering system, now with the ability to pass a
parameter to a filter function (filters get setup functions as a
result). This also included adding n00b_flush() particualarly to try
to give a path for getting something useful out of partial filter
state.

Ported the tiny filters (e.g. for line buffering). Also did the hex
dump and ansi filter, which were more substantial.

Some linux porting work.

Along the way, I added an option to protect the entire heap, and align
allocs to the page, so if there's a pointer math error, I can detect
it as it happens.