Hex dump filter
Added filters for line buffering and hex dumps. Also added the ability to pass a parameter to a filter function (filters get setup functions as a result). This also included adding n00b_flush() particualarly to try to give a path for getting something useful out of partial filter state.