feat: Debugging APIs
Re-built the debug server on the new system.  Can support server-side
filtering as a result.

Added in the test client a subscription to server messages so that
closes will propogate to us.

Also, apparently Linux TSI destructors remove your key before the
deletion callback?? Had to revive the thread-level clean-up handler,
which seems to reliably run before the TSI destructor.