fix: compilation error on linux
n00b_local_process_server in workflow.h wasn't defined as an extern but
was referenced and its value set in server.c, which was causing "multiple
definition of `n00b_local_process_is_server'" errors when trying to compile
on Linux. This fixes the compilation errors by declaring it as an extern, but
there's still some instability when running ./build_dev/io_demo on Linux
that manifests as io_demo unpredictably hanging.