fix: Linux portability
Did the work to get things running on Linux reliably again.

WOW, pthread_getattr_np() on Linux is super-duper slow. I was calling
it any time a thread 'suspends', where it wasn't necessary (it can be
called once at the beginning of a thread's life). Still, I'd expect it
to just copy in a couple of pointers, but it is super hard-core;
calling it that often was basically slowing things down by 200x or
more.  Mind blown.

Also involved some icky hackiness to avoid having to encode info
specific to the va_args implementation. But this got really ugly,
because I don't seem to be able to portably cast my way out of all the
warnings these macros generate on their way to not having to deal w/
va_list implementation details on a per-stdlib basis.