This *adds* [= =] as valid digraph syntax (in addition to the current [| |] syntax).
The current syntax is a bit more aesthetic and wouldn't expect it to
be a reasonable character combo in string literals (which is
desirable). But [= =] is much easier to type, and almost as uncommon
(its downside is that it looks too much like an ASCII smile, even
though that is pretty unused in the modern world, so seems
reasonable).

Would be happy to consider other options, but after trying out [| |]
for quite a white, I can say I still frequently find it difficult to
type (it's a combination of switching between 'shift' and the fact
that ] and | are right next to each other).

Seems like having both is a reasonable compromise for the moment.

Still, I converted all current instances of [| ... |] to the unicode «
» syntax, except in the testing where appropriate.