Finally getting around to making the type system changes I have been wanting to make.
Two main issues:

1) I wanted concrete runtime types to be just an int, not trees. This
   wasn't having an impact when profiling, but still felt icky and
   wanted to fix.

2) The type hashing was occasionally problematic because types are
   used early, and the hashing algorithm did require dynamic allocation.

The second one occasionally led to issues when I mucked around with
the startup (which I was doing, which is what motivated me doing this
now).

In this new implementation, the hashing algorithm only is applied to
concrete types when instantiating type trees for the purposes of
comparison, and no longer will require dynamic allocation.

The enumeration of builtin primitive base types base been reordered;
concrete (non-parameterized) base types all must come first. That's
because the 64-bit type_id is no longer a pointer, it is an
integer. The integer is constructed based on the type:

1) If the type is parameterized, but it is CONCRETE, then the type is
   a 63-bit hash of the type tree. For that, I'm using Komihash, and
   masking out the most significant bit (which distinguishes from
   non-hashed types).

2) Otherwise, types are offsets into slots in the type store, which is
   now a single, contiguous, seprately managed section of memory.  The
   type slots for all primitive types are always the builtin type ID.

Note that this approach will end up requiring recompliation of things
whenever the number of builtin concrete types change. However, once I
add the object system, I'm hoping this number will only ever go down.

Additionally, while you can still have multiple type "universes", the
API is now much simpler due to the keywords system-- instead of 2
APIs, there's just one, and if you want to use anything other than the
default universe, you pass your universe via a keyword parameter.

The core of the new system is in place. I need to add a few of the
more more advanced features, but the big lift from here is going to be
migrating the rest of the code.